Fix linter error
