+++ modules/book.js
@@ -1,23 +1,9 @@
-import { Collect } from './collect.js';
 const inputTitle = document.getElementById('title');
 const inputAuthor = document.getElementById('author');
 const submitBtn = document.querySelector('.add-btn');
-class Book {
+export class Book {
   constructor(title, author) {
     this.title = title;
     this.author = author;
   }
 }
-
-const coll = new Collect();
-if (localStorage.getItem('bookCollection')) {
-  const localBooks = JSON.parse(localStorage.getItem('bookCollection'));
-  localBooks.bookColl.forEach((element) => {
-    coll.add(new Book(element.title, element.author));
-  });
-}
-submitBtn.addEventListener('click', () => {
-  coll.add(new Book(inputTitle.value, inputAuthor.value));
-});
-
-export { Book };
