+++ modules/collect.js
@@ -1,5 +1,5 @@
 //import { populateStorage } from './populate.js';
-
+import { Book } from './book.js';
 const inputTitle = document.getElementById('title');
 const inputAuthor = document.getElementById('author');
 const submitBtn = document.querySelector('.add-btn');
@@ -52,4 +52,14 @@ class Collect {
     );
   };
 }
+const coll = new Collect();
+if (localStorage.getItem('bookCollection')) {
+  const localBooks = JSON.parse(localStorage.getItem('bookCollection'));
+  localBooks.bookColl.forEach((element) => {
+    coll.add(new Book(element.title, element.author));
+  });
+}
+submitBtn.addEventListener('click', () => {
+  coll.add(new Book(inputTitle.value, inputAuthor.value));
+});
 export { Collect };
